chore!: rename top-level lib folder to `open_feature`
BREAKING CHANGE: Since we've decided on `OpenFeature` as our top-level module name, we should follow the Ruby convention of separating the words with an underscore in the file name.
