+++ .github/workflows/release-please.yml
@@ -16,7 +16,7 @@ jobs:
           release-type: ruby
           package-name: openfeature-sdk
           bump-minor-pre-major: true
-          version-file: "lib/openfeature/sdk/version.rb"
+          version-file: "lib/open_feature/sdk/version.rb"
     outputs:
       release_created: ${{ steps.release.outputs.release_created }}
       release_tag_name: ${{ steps.release.outputs.tag_name }}
