+++ lib/open_feature/sdk/client.rb
