+++ lib/open_feature/sdk/configuration.rb
