+++ lib/open_feature/sdk/metadata.rb
