+++ lib/open_feature/sdk/provider/no_op_provider.rb
